Increased game functionality
Tried to do a couple things today. I succesfully combined two functions and dispatched them correctly making the game a lot more accessible if I can put it that way. Also added a reference to the textarea so allow focusing after pressing button to switch/initialize the quote. I tried to disable the input box after the text area and the quote had the same length but wasnt able to get the reducer working with the boolean so I may have to make another reducer to update the state but that shouldnt be neccessary